fix: drive backend to a clean lint/type/deprecation gate (ruff + mypy + no warnings) (#19)
System-test cleanup pass across the Bid_it backend, run through an
architect→developer→engineer→tester loop and independently verified.

- Deprecation (B2): rename every status.HTTP_422_UNPROCESSABLE_ENTITY to
  HTTP_422_UNPROCESSABLE_CONTENT (identical 422 value) across ~15 route/service
  files — clears the 28 StarletteDeprecationWarnings from the FastAPI 0.139 bump
  and future-proofs against the constant's removal.
- Latent None→500 + type safety (B3/B4): add a get_current_org dependency
  (CurrentOrg) that resolves the caller's org once and raises 401 if it is
  missing, replacing unguarded `await db.get(Organization, ...)` + attribute
  access in the billing/settings/etc. routes. Remaining mypy gaps closed with
  targeted None-guards, typed casts of SQLAlchemy CursorResult.rowcount, and
  invariant asserts (each documented) — no `# type: ignore` suppressions.
- Observability (B5): log inside previously-silent best-effort except handlers
  (invoice_pdf logo render, einvoice attachment probe, observability hooks) —
  control flow unchanged, failures now visible.
- Lint (B1): ruff import-order fix in alembic/env.py.

Verified: ruff clean, mypy clean (229 files), frontend tsc clean, vite build
clean, full backend suite green (761 passed, 1 skipped). Behavior preserved;
the two touched test files are formatting-only (no assertions changed).

+++ backend/app/models/expense_approval.py
@@ -13,6 +13,7 @@ Tenant isolation is via the denormalised `org_id` + the ORM guard + Postgres RLS
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     Boolean,
@@ -55,7 +56,7 @@ class ExpenseApprovalPolicy(UUIDPrimaryKeyMixin, TimestampMixin, Base):
     # Lower runs first; the first fully-matching active policy wins.
     priority: Mapped[int] = mapped_column(Integer, default=100, nullable=False)
     # Match criterion (null = wildcard): the report's EUR total must be >= this.
-    min_amount: Mapped[float | None] = mapped_column(Numeric(14, 2), nullable=True)
+    min_amount: Mapped[Decimal | None] = mapped_column(Numeric(14, 2), nullable=True)
     # Ordered chain: a JSON array of user-ids, each an approval step in order.
     approver_ids: Mapped[str | None] = mapped_column(Text, nullable=True)
     # Append a finance sign-off step at the end of the chain.
